requirements.txt
pandas
sqlalchemy
psycopg2-binary
boto3
prefect
python-dotenv
scikit-learn
numpy